Refactor dir structure and cleanup.
