+++ contracts/erc2o/test/contracts/eosio.token/eosio.token.abi
