+++ contracts/erc2o/test/main.cpp
@@ -8,9 +8,7 @@
 #include <boost/test/included/unit_test.hpp>
 #include <fc/log/logger.hpp>
 #include <eosio/chain/exceptions.hpp>
-#include "eosio.system_tester.hpp"
 
-using namespace eosio_system;
 #define BOOST_TEST_STATIC_LINK
 
 void translate_fc_exception(const fc::exception &e) {
