--- test/contracts/eosio.wrap/eosio.wrap.abi
@@ -1,130 +1,0 @@
-{
-    "____comment": "This file was generated with eosio-abigen. DO NOT EDIT Fri Nov 16 11:30:52 2018",
-    "version": "eosio::abi/1.1",
-    "structs": [
-        {
-            "name": "action",
-            "base": "",
-            "fields": [
-                {
-                    "name": "account",
-                    "type": "name"
-                },
-                {
-                    "name": "name",
-                    "type": "name"
-                },
-                {
-                    "name": "authorization",
-                    "type": "permission_level[]"
-                },
-                {
-                    "name": "data",
-                    "type": "bytes"
-                }
-            ]
-        },
-        {
-            "name": "exec",
-            "base": "",
-            "fields": [
-                {
-                    "name": "executer",
-                    "type": "name"
-                },
-                {
-                    "name": "trx",
-                    "type": "transaction"
-                }
-            ]
-        },
-        {
-            "name": "extension",
-            "base": "",
-            "fields": [
-                {
-                    "name": "type",
-                    "type": "uint16"
-                },
-                {
-                    "name": "data",
-                    "type": "bytes"
-                }
-            ]
-        },
-        {
-            "name": "permission_level",
-            "base": "",
-            "fields": [
-                {
-                    "name": "actor",
-                    "type": "name"
-                },
-                {
-                    "name": "permission",
-                    "type": "name"
-                }
-            ]
-        },
-        {
-            "name": "transaction",
-            "base": "transaction_header",
-            "fields": [
-                {
-                    "name": "context_free_actions",
-                    "type": "action[]"
-                },
-                {
-                    "name": "actions",
-                    "type": "action[]"
-                },
-                {
-                    "name": "transaction_extensions",
-                    "type": "extension[]"
-                }
-            ]
-        },
-        {
-            "name": "transaction_header",
-            "base": "",
-            "fields": [
-                {
-                    "name": "expiration",
-                    "type": "time_point_sec"
-                },
-                {
-                    "name": "ref_block_num",
-                    "type": "uint16"
-                },
-                {
-                    "name": "ref_block_prefix",
-                    "type": "uint32"
-                },
-                {
-                    "name": "max_net_usage_words",
-                    "type": "varuint32"
-                },
-                {
-                    "name": "max_cpu_usage_ms",
-                    "type": "uint8"
-                },
-                {
-                    "name": "delay_sec",
-                    "type": "varuint32"
-                }
-            ]
-        }
-    ],
-    "types": [],
-    "actions": [
-        {
-            "name": "exec",
-            "type": "exec",
-            "ricardian_contract": ""
-        }
-    ],
-    "tables": [],
-    "ricardian_clauses": [],
-    "variants": [],
-    "abi_extensions": []
-}
